Move fixing updates. Replace end of lines.
