+++ can/interfaces/ixxat/constants.py
