+++ can/interfaces/ixxat/exceptions.py
